services/users/requirements.txt
coverage==4.5.1
flake8==3.5.0
Flask==1.0.2
flask-bcrypt==0.7.1
flask-cors==3.0.6
flask-debugtoolbar==0.10.1
flask-migrate==2.2.0
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.6.2
Gunicorn==19.8.1
psycopg2-binary==2.7.5
pyjwt==1.6.4